requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
apscheduler==3.10.4
aiohttp==3.9.5
